+++ server/test/unit/apiTest/authorizeTest.js
@@ -6,29 +6,13 @@ const sinonChai = require('sinon-chai')
 
 // In-house dependencies
 const { helpers } = require('brave-alert-lib')
-
+const { mockResponse } = require('../../../testingHelpers')
 const api = require('../../../api')
 
 // Configure Chai
 use(sinonChai)
 
 const sandbox = sinon.createSandbox()
-
-function authorize_next() {
-  helpers.log('authorize_next')
-}
-
-function mockRequest(_req) {
-  const req = _req
-  return req
-}
-
-function mockResponse() {
-  const res = {}
-  res.status = sinon.stub().returns(res)
-  res.send = sinon.stub().returns(res)
-  return res
-}
 
 describe('api.js unit tests: authorize', () => {
   beforeEach(() => {
@@ -44,156 +28,184 @@ describe('api.js unit tests: authorize', () => {
   describe('when a request provides the primary PA API key in the query string of the request', () => {
     beforeEach(async () => {
       // create fake request object that implements the necessary members
-      const req = mockRequest({
+      const req = {
         query: { braveKey: helpers.getEnvVar('PA_API_KEY_PRIMARY') },
         body: {}, // empty
         path: 'test_path',
-      })
-      this.res = mockResponse()
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
 
-      await api.authorize(req, this.res, authorize_next)
+      await api.authorize(req, this.res, this.next)
     })
 
-    it('should call the authorize_next function', () => {
-      expect(helpers.log).to.have.been.calledWith('authorize_next')
+    it('should call the next function', () => {
+      expect(this.next).to.be.called
+    })
+
+    it('should not set the response status', () => {
+      expect(this.res.status).to.not.be.called
     })
   })
 
   describe('when a request provides the primary PA API key in the body of the request', () => {
     beforeEach(async () => {
       // create fake request object that implements the necessary members
-      const req = mockRequest({
+      const req = {
         query: {}, // empty
         body: { braveKey: helpers.getEnvVar('PA_API_KEY_PRIMARY') },
         path: 'test_path',
-      })
-      this.res = mockResponse()
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
 
-      await api.authorize(req, this.res, authorize_next)
+      await api.authorize(req, this.res, this.next)
     })
 
-    it('should call the authorize_next function', () => {
-      expect(helpers.log).to.have.been.calledWith('authorize_next')
+    it('should call the next function', () => {
+      expect(this.next).to.be.called
+    })
+
+    it('should not set the response status', () => {
+      expect(this.res.status).to.not.be.called
     })
   })
 
   describe('when a request provides the secondary PA API key in the query string of the request', () => {
     beforeEach(async () => {
       // create fake request object that implements the necessary members
-      const req = mockRequest({
+      const req = {
         query: { braveKey: helpers.getEnvVar('PA_API_KEY_SECONDARY') },
         body: {}, // empty
         path: 'test_path',
-      })
-      this.res = mockResponse()
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
 
-      await api.authorize(req, this.res, authorize_next)
+      await api.authorize(req, this.res, this.next)
     })
 
-    it('should call the authorize_next function', () => {
-      expect(helpers.log).to.have.been.calledWith('authorize_next')
-    })
-  })
-
-  describe('when a request provides the secondary PA API key in the body of the request', () => {
-    beforeEach(async () => {
-      // create fake request object that implements the necessary members
-      const req = mockRequest({
-        query: {}, // empty
-        body: { braveKey: helpers.getEnvVar('PA_API_KEY_SECONDARY') },
-        path: 'test_path',
-      })
-      this.res = mockResponse()
-
-      await api.authorize(req, this.res, authorize_next)
+    it('should call the next function', () => {
+      expect(this.next).to.be.called
     })
 
-    it('should call the authorize_next function', () => {
-      expect(helpers.log).to.have.been.calledWith('authorize_next')
-    })
-  })
-
-  describe('when a request provides an invalid PA API key in the query string of the request', () => {
-    beforeEach(async () => {
-      // create fake request object that implements the necessary members
-      const req = mockRequest({
-        query: { braveKey: 'INVALID_API_KEY' },
-        body: {}, // empty
-        path: 'test_path',
-      })
-      this.res = mockResponse()
-
-      await api.authorize(req, this.res, authorize_next)
-    })
-
-    it('should not call the authorize_next function', () => {
-      expect(helpers.log).to.not.have.been.calledWith('authorize_next')
-    })
-
-    it('should result in status 401 (Unauthorized)', () => {
-      expect(this.res.status).to.have.been.calledWith(401)
-    })
-  })
-
-  describe('when a request provides an invalid PA API key in the body of the request', () => {
-    beforeEach(async () => {
-      // create fake request object that implements the necessary members
-      const req = mockRequest({
-        query: {}, // empty
-        body: { braveKey: 'INVALID_API_KEY' },
-        path: 'test_path',
-      })
-      this.res = mockResponse()
-
-      await api.authorize(req, this.res, authorize_next)
-    })
-
-    it('should not call the authorize_next function', () => {
-      expect(helpers.log).to.not.have.been.calledWith('authorize_next')
-    })
-
-    it('should result in status 401 (Unauthorized)', () => {
-      expect(this.res.status).to.have.been.calledWith(401)
-    })
-  })
-
-  describe('when a request does not provide a PA API key at all', () => {
-    beforeEach(async () => {
-      // create fake request object that implements the necessary members
-      const req = mockRequest({
-        query: {}, // empty
-        body: {}, // empty
-        path: 'test_path',
-      })
-      this.res = mockResponse()
-
-      await api.authorize(req, this.res, authorize_next)
-    })
-
-    it('should not call the authorize_next function', () => {
-      expect(helpers.log).to.not.have.been.calledWith('authorize_next')
-    })
-
-    it('should result in status 401 (Unauthorized)', () => {
-      expect(this.res.status).to.have.been.calledWith(401)
+    it('should not set the response status', () => {
+      expect(this.res.status).to.not.be.called
     })
   })
 
   describe('when a request provides the primary PA API key in both the query and the body', () => {
     beforeEach(async () => {
       // create fake request object that implements the necessary members
-      const req = mockRequest({
+      const req = {
         query: { braveKey: helpers.getEnvVar('PA_API_KEY_PRIMARY') },
         body: { braveKey: helpers.getEnvVar('PA_API_KEY_PRIMARY') },
         path: 'test_path',
-      })
-      this.res = mockResponse()
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
 
-      await api.authorize(req, this.res, authorize_next)
+      await api.authorize(req, this.res, this.next)
     })
 
-    it('should call the authorize_next function', () => {
-      expect(helpers.log).to.have.been.calledWith('authorize_next')
+    it('should call the next function', () => {
+      expect(this.next).to.be.called
+    })
+
+    it('should not set the response status', () => {
+      expect(this.res.status).to.not.be.called
+    })
+  })
+
+  describe('when a request provides the secondary PA API key in the body of the request', () => {
+    beforeEach(async () => {
+      // create fake request object that implements the necessary members
+      const req = {
+        query: {}, // empty
+        body: { braveKey: helpers.getEnvVar('PA_API_KEY_SECONDARY') },
+        path: 'test_path',
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
+
+      await api.authorize(req, this.res, this.next)
+    })
+
+    it('should call the next function', () => {
+      expect(this.next).to.be.called
+    })
+
+    it('should not set the response status', () => {
+      expect(this.res.status).to.not.be.called
+    })
+  })
+
+  describe('when a request provides an invalid PA API key in the query string of the request', () => {
+    beforeEach(async () => {
+      // create fake request object that implements the necessary members
+      const req = {
+        query: { braveKey: 'INVALID_API_KEY' },
+        body: {}, // empty
+        path: 'test_path',
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
+
+      await api.authorize(req, this.res, this.next)
+    })
+
+    it('should not call the next function', () => {
+      expect(this.next).to.not.be.called
+    })
+
+    it('should respond with status 401 (Unauthorized)', () => {
+      expect(this.res.status).to.be.calledWith(401)
+    })
+  })
+
+  describe('when a request provides an invalid PA API key in the body of the request', () => {
+    beforeEach(async () => {
+      // create fake request object that implements the necessary members
+      const req = {
+        query: {}, // empty
+        body: { braveKey: 'INVALID_API_KEY' },
+        path: 'test_path',
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
+
+      await api.authorize(req, this.res, this.next)
+    })
+
+    it('should not call the next function', () => {
+      expect(this.next).to.not.be.called
+    })
+
+    it('should respond with status 401 (Unauthorized)', () => {
+      expect(this.res.status).to.be.calledWith(401)
+    })
+  })
+
+  describe('when a request does not provide a PA API key at all', () => {
+    beforeEach(async () => {
+      // create fake request object that implements the necessary members
+      const req = {
+        query: {}, // empty
+        body: {}, // empty
+        path: 'test_path',
+      }
+      this.res = mockResponse(sandbox)
+      this.next = sandbox.stub()
+
+      await api.authorize(req, this.res, this.next)
+    })
+
+    it('should not call the next function', () => {
+      expect(this.next).to.not.be.called
+    })
+
+    it('should respond with status 401 (Unauthorized)', () => {
+      expect(this.res.status).to.be.calledWith(401)
     })
   })
 })